+++ src/config/defaultsOptions.js
@@ -1,10 +1,10 @@
 export const defaultsOptions = Object.freeze({
   prefix: 'socket_',
   ERROR_EVENTS: ['error', 'api-error'],
-  errorHandler: ({ error, logic, input, socket } = {}) => {
+  errorHandler: ({ error, getKeaContext, socket } = {}) => {
     console.error('[kea-socket.io] ' + error);
     console.error(socket);
-    console.error(logic);
+    console.error(getKeaContext());
   },
   mapSocketEventToStore: ({ name }) => name,
   emitterActions: {}