+++ CHANGELOG.md
@@ -6,6 +6,13 @@ As we're at the 0.x phase, deprecations and breaking changes will still happen. 
 
 Once we react 1.0 all deprecations will be removed and the project will switch to SemVer.
 
+## 0.3.0 - 2020-03-01
+
+### Changed
+
+- Rebuild all observe inside logic (nothing change in interfaces)
+- Change error handler interface
+
 ## 0.2.4 - 2020-02-29
 
 ### Changed